+++ src/aleph/chains/cosmos.py
@@ -10,7 +10,7 @@ from ecdsa import BadSignatureError
 
 from aleph.chains.common import get_verification_buffer
 from aleph.schemas.pending_messages import BasePendingMessage
-from .connector import Verifier
+from .abc import Verifier
 
 LOGGER = logging.getLogger("chains.cosmos")
 CHAIN_NAME = "CSDK"
