+++ src/aleph/chains/solana.py
@@ -7,7 +7,7 @@ from nacl.signing import VerifyKey
 
 from aleph.chains.common import get_verification_buffer
 from aleph.schemas.pending_messages import BasePendingMessage
-from .connector import Verifier
+from .abc import Verifier
 
 LOGGER = logging.getLogger("chains.solana")
 CHAIN_NAME = "SOL"
