+++ src/aleph/chains/substrate.py
@@ -5,7 +5,7 @@ from substrateinterface import Keypair
 
 from aleph.chains.common import get_verification_buffer
 from aleph.schemas.pending_messages import BasePendingMessage
-from .connector import Verifier
+from .abc import Verifier
 
 LOGGER = logging.getLogger("chains.substrate")
