+++ tests/api/conftest.py
@@ -1,3 +1,4 @@
+import datetime as dt
 import json
 from pathlib import Path
 from typing import Any, Dict, Sequence, cast, Tuple
@@ -8,7 +9,7 @@ from aleph_message.models import AggregateContent, PostContent
 from configmanager import Config
 from sqlalchemy import insert
 
-from aleph.chains.chain_service import ChainService
+from aleph.chains.signature_verifier import SignatureVerifier
 from aleph.db.accessors.aggregates import refresh_aggregate
 from aleph.db.models import (
     MessageDb,
@@ -22,8 +23,6 @@ from aleph.jobs.process_pending_messages import PendingMessageProcessor
 from aleph.storage import StorageService
 from aleph.toolkit.timestamp import timestamp_to_datetime
 from aleph.types.db_session import DbSessionFactory
-import datetime as dt
-
 from in_memory_storage_engine import InMemoryStorageEngine
 
 
@@ -205,12 +204,10 @@ def message_processor(mocker, mock_config: Config, session_factory: DbSessionFac
         ipfs_service=mocker.AsyncMock(),
         node_cache=mocker.AsyncMock(),
     )
-    chain_service = ChainService(
-        session_factory=session_factory, storage_service=storage_service
-    )
+    signature_verifier = SignatureVerifier()
     message_handler = MessageHandler(
         session_factory=session_factory,
-        chain_service=chain_service,
+        signature_verifier=signature_verifier,
         storage_service=storage_service,
         config=mock_config,
     )
