+++ tests/message_processing/test_process_pending_txs.py
@@ -7,6 +7,7 @@ from aleph_message.models import Chain, MessageType, PostContent
 from configmanager import Config
 from sqlalchemy import select
 
+from aleph.chains.signature_verifier import SignatureVerifier
 from aleph.chains.chaindata import ChainDataService
 from aleph.db.models import PendingMessageDb, MessageStatusDb
 from aleph.db.models.chains import ChainTxDb
@@ -39,12 +40,13 @@ async def test_process_pending_tx_on_chain_protocol(
 ):
     chain_data_service = mocker.AsyncMock()
     chain_data_service.get_tx_messages = get_fixture_chaindata_messages
+    signature_verifier = SignatureVerifier()
     pending_tx_processor = PendingTxProcessor(
         session_factory=session_factory,
         storage_service=test_storage_service,
         message_handler=MessageHandler(
             session_factory=session_factory,
-            chain_service=mocker.AsyncMock(),
+            signature_verifier=signature_verifier,
             storage_service=test_storage_service,
             config=mock_config,
         ),
@@ -113,12 +115,13 @@ async def _process_smart_contract_tx(
     chain_data_service = ChainDataService(
         session_factory=session_factory, storage_service=mocker.AsyncMock()
     )
+    signature_verifier = SignatureVerifier()
     pending_tx_processor = PendingTxProcessor(
         session_factory=session_factory,
         storage_service=test_storage_service,
         message_handler=MessageHandler(
             session_factory=session_factory,
-            chain_service=mocker.AsyncMock(),
+            signature_verifier=signature_verifier,
             storage_service=test_storage_service,
             config=mock_config,
         ),
