+++ tests/test_network.py
@@ -3,7 +3,8 @@ import datetime as dt
 import pytest
 from configmanager import Config
 
-from aleph.chains.chain_service import ChainService
+from aleph.chains.connector import ChainConnector
+from aleph.chains.signature_verifier import SignatureVerifier
 from aleph.db.models import PendingMessageDb
 from aleph.handlers.message_handler import MessageHandler
 from aleph.schemas.pending_messages import parse_message
@@ -26,11 +27,9 @@ async def test_valid_message(mocker):
         "signature": "2103041b0b357446927d2c8c62fdddd27910d82f665f16a4907a2be927b5901f5e6c004730450221009a54ecaff6869664e94ad68554520c79c21d4f63822864bd910f9916c32c1b5602201576053180d225ec173fb0b6e4af5efb2dc474ce6aa77a3bdd67fd14e1d806b4",
     }
 
-    chain_service = ChainService(
-        session_factory=mocker.AsyncMock(), storage_service=mocker.MagicMock()
-    )
+    signature_verifier = SignatureVerifier()
     sample_message = parse_message(sample_message_dict)
-    await chain_service.verify_signature(sample_message)
+    await signature_verifier.verify_signature(sample_message)
 
 
 @pytest.mark.asyncio
@@ -63,13 +62,11 @@ async def test_invalid_signature_message(mocker):
         "signature": "BAR",
     }
 
-    chain_service = ChainService(
-        session_factory=mocker.AsyncMock(), storage_service=mocker.MagicMock()
-    )
+    signature_verifier = SignatureVerifier()
 
     sample_message = parse_message(sample_message_dict)
     with pytest.raises(InvalidMessageException):
-        _ = await chain_service.verify_signature(sample_message)
+        _ = await signature_verifier.verify_signature(sample_message)
 
 
 @pytest.mark.asyncio
@@ -84,13 +81,12 @@ async def test_invalid_signature_message_2(mocker):
         "time": 1563279102.3155158,
         "signature": "2153041b0b357446927d2c8c62fdddd27910d82f665f16a4907a2be927b5901f5e6c004730450221009a54ecaff6869664e94ad68554525c79c21d4f63822864bd910f9916c32c1b5602201576053180d225ec173fb0b6e4af5efb2dc474ce6aa77a3bdd67fd14e1d806b4",
     }
-    chain_service = ChainService(
-        session_factory=mocker.AsyncMock(), storage_service=mocker.MagicMock()
-    )
+
+    signature_verifier = SignatureVerifier()
 
     sample_message = parse_message(sample_message_dict)
     with pytest.raises(InvalidMessageException):
-        _ = await chain_service.verify_signature(sample_message)
+        _ = await signature_verifier.verify_signature(sample_message)
 
 
 @pytest.mark.asyncio
@@ -111,11 +107,10 @@ async def test_incoming_inline_content(
         "signature": "21027c108022f992f090bbe5c78ca8822f5b7adceb705ae2cd5318543d7bcdd2a74700473045022100b59f7df5333d57080a93be53b9af74e66a284170ec493455e675eb2539ac21db022077ffc66fe8dde7707038344496a85266bf42af1240017d4e1fa0d7068c588ca7",
     }
 
+    signature_verifier = SignatureVerifier()
     message_handler = MessageHandler(
         session_factory=session_factory,
-        chain_service=ChainService(
-            session_factory=session_factory, storage_service=test_storage_service
-        ),
+        signature_verifier=signature_verifier,
         storage_service=test_storage_service,
         config=mock_config,
     )
